I think this works

+++ 13Heaps/Driver.java
@@ -1,7 +1,9 @@
+import java.util.*;
+
 public class Driver{
 
     public static void main(String[]args){
-	MyHeap<Integer> test = new MyHeap<Integer>(false);
+	MyHeap<Integer> test = new MyHeap<Integer>();
 	System.out.println(test);
 	test.add(1);
 	System.out.println(test);
@@ -15,7 +17,24 @@ public class Driver{
 	System.out.println(test);
 	test.add(2);
 	System.out.println(test);
-	test.delete();
+	try{
+	    test.delete();
+	}catch(Exception e){
+	    System.out.println("Working");
+	}
+	System.out.println(test);
+	Random r = new Random();	
+	for(int i=0;i<1000;i++){
+	    test.add(r.nextInt(10)*(r.nextInt(3)-1));
+	    if(r.nextInt(2)==0){
+		try{
+		    test.delete();
+		}catch(Exception e){
+		    
+		}
+	    }
+	}
+	
 	System.out.println(test);
 	System.out.println(test.peek());
     }
